requirements.txt
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
PyJWT==2.8.0
gunicorn==22.0.0

# Uncomment for Azure deployment:
# PyMySQL==1.1.1
# azure-storage-blob==12.20.0

# Uncomment for GCP deployment:
# google-cloud-storage==2.17.0
